Example to demonstrate multi threading in Arduino
This example uses Timer1 in Arduino to generate intervals at preset time
and take actions (blink the on-board LED in one function and print
"Hello" to Serial port in another function). It does not use delay()
function which is a big minus point as the main loop halts and prevents
other actions from execution.